Add "so as to" to simplify rules
I'm fairly sure you'll agree with this addition, but let me know if you don't.

I'm TAing for a writing course and one of my students used "so as to" a lot where they could have just used "to". I planned on using that mistake as an opportunity to point them to this style guide, but I realized that you didn't actually have that specific example.